Refactor methods for SML Classes generation
As the basyx sdk uses idShort hack for SMLists by setting
some generated idShort for SMElements in the SML,
we had to fix the issue by resetting the idShort to smth,
what makes more sense, to generate the SMElement classes.
In the end we remove the idShort and set it to None, as
stated in the spec.